Refacted code into utility functions
